WIP: implement comparator inventory measurement
Detect container at the back side of the comparator. The container may be
separated from the comparator by an opaque block.

If an inventory is detected the power state depends on the fullness of the
container from 0 (all slots empty) to 15 (all slots contain a full stack).

Chest inventory actions will trigger an update of comparators in range.