Pathfinding improvements
* mods/ENTITIES/mcl_mobs/pathfinding.lua (gwp_reconstruct):
Return node of arrival in candidate.target.
(gwp_skip_waypoint): Also skip waypoint if the difference
between the horizontal directions from the previous surface to
the current position and from the current position to the next
is smaller than 15 degrees, as when falling from one waypoint to
the next.
(gwp_next_waypoint): Delegate skip tests to gwp_skip_waypoint.