Turn craftguide progress list into a set
This speeds up filtering significantly for an increase in initialization
time: each item has to be tested whether it fulfills any of the relevant
group specifications from the group_cache.

Seen items are still stored as a list in player metadata for backwards
compatibility (and also because the serialized form is more compact).