Rework fence interaction
Move all logic to leads, override group and on_rightclick of fence nodes after
mods loaded. All changes to mcl_fences are removed.

Clicking a knot will now behave identically to clicking the fence: first try to
take a lead (can be skipped by holding sneak), then try to tie a lead to the
fence and last try to tie a new lead to the fence.

Fix taking the follower side of a lead as well as trying to tie a fence to
itself.

Uses the _mcl_on_rightclick_optional feature to inject lead handling when
rightclicking fences while not wielding a lead item.

Also refactor check of can_lead_attach group