Fix lit [exposed] copper bulb texture mipmaps
Luanti 5.10 displayed corrupted textures for both lit copper bulbs and
lit exposed copper bulbs, if mipmaps were enabled. This patch replaces
both affected textures with ones that have no alpha channel. For this,
the mtPaint 3.40 “remove channel” functionality was used; removing the
transparency using the GIMP 2.10 did not produce the correct textures.

It seems that the alpha channel was not necessary in the first place –
as lit copper bulb nodes do not have any visible transparency in-game.