Ignore pushing [sticky] piston position when calculating what is pushed
When pushing several nodes using a [sticky] piston that pushed a sticky
(slime or honey) block, the pushing [sticky] piston was counted towards
the push limit. This resulted in pistons being able to push fewer nodes
than intended (eleven instead of twelve). Pulling nodes was unaffected.

This patch makes the code skip the pushing [sticky] piston position for
purposes of calculating which nodes are pushed, thereby fixing the bug.