Optimization: Avoid redundant calls to minetest.get_objects_inside_radius
When not pushing upwards, moved blocks positions as well as positions above moved block are added to `move_positions` (or handled right away in the old implementation). This can result in unnecessary calls to `minetest.get_objects_inside_radius`, for example when pushing vertically stacked slime blocks sideways.

Added a table to avoid processing duplicate positions. An alternative is to do it before insertion.