Simplify set_pending_update a bit
(Small follow-up change related to PR 3052). Removes the behavior/"feature" where internal use of set_pending_update on a position without an update callback would remove the position by setting it to nil. Check for update callback is done when processing mcl_redstone._pending_updates, so the only side-effect is that the position stays in the table until updates are processed.

A benefit of not doing "set to nil" is that it eliminates the requirement of implementing "remove(value)" if wanting to change the underlying data structure, e.g. to one that preserves insertion order.