Add more facedir and new wallmounted attachment groups
The main difference between them is what helper function is used to
figure out the direction to figure out the position to check for.
Further improvements may include checking for the attached node's
`paramtype2` in order to figure that out, but I decided on continuing
the pattern for now and to have this be compatible with VL groups.