adjust snow fall to happen even on opaque blocks
The check of the natural light of a block in
mcl_weather.can_see_outdoors() probably returns 0 on opaque
blocks, because before this commit snow will never start
accumulating during a snow storm on a bare block of dirt. Snow
can accumulate on leaves, however, which pass some light
through. This patch checks the natural light halfway up this node
which will more correctly determine light for the purpose of
piling up snow.