Fix single vs. double pulse in 2-tick activation pulse scenarios.
Short story is that when observer notifications are spaced exactly 2 ticks apart it's kind of a grey area what should happen - i.e. it depends on in which order events unfold within a single tick.

The observer implementation is designed around the premise that non-scheduled updates/events are invoked immediately in a depth first fashion, which isn't the case when it comes to the redstone implementation. The result is that effect (e.g deactivate observer) is sometimes scheduled before cause (e.g. turn off repeater).