Ugly workaround for enum player setting tooltip
When an open dropdown list without a tooltip overlays an area with a
tooltip that tooltip will be shown when mousing over the dropdown options.

Therefore set an unobtrusive tooltip for the dropdown overriding any area
tooltip. This is ugly, but still a bit better than showing a large tooltip
with the enum longdesc when trying to select an option with the mouse.

Ultimately this needs to be solved in the engine by e.g. not showing any
area tooltip when mousing over an open dropdown list.