Really limit size of stack moved into armor slot to 1
Currently it is possible to put a multi item stack into an empty armor slot.

Extends the condition that prevents replacing an armor item with a non
singular stack to always limit the amount of moved items to 1.

Also adds some parentheses to the condition, because it seems that unintended
operator precedence was the root cause of the problem.