Launch tridents from dispensers as in MC Bedrock Edition &c.
* mods/ITEMS/mcl_tridents/init.lua (_on_dispense, is_walkable):
New functions.
(on_step): Reproduce arrow mechanics as respects redstone;
tridents are internally variants of arrows in Minecraft.  Do not
step loyalty timer if the shooter is not online or restore
regular movement if the node in which this trident is embedded
is walkable.
(shoot_trident): New arg SPEED.  All callers changed.  Also,
handle scenarios where OBJ is nil.