Fix lilac patches in singlenode mapgen
The patch_lilac flower feature incorrectly uses blue orchids as its block_content, causing them to spawn instead of lilacs in worlds that use the new map generation. This commit fixes that behavior.

How to test:
	1- Find a biome where lilac patches can spawn, such as forests
	2- Check that lilacs properly spawn, instead of blue orchids

Both screenshots are on seed 1612; it turns out that singlenode mapgen's a bit too much for my computer, so I had a few difficulties checking that this properly works. (Pay no mind to my inventory in the first screenshot; I was testing out a mod I'm making)