Remove stale crafting registration.
Mcl_amethyst still has a crafting recipe for the spyglass tool,
with conditionals testing for the presence of the spyglass and the
copper mod.

Since both mods have for a while been an integral part of the game,
and the spyglass mod already has the correct crafting registration,
this stale code can be removed.